# docs/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Fellow Travelers API
  version: 0.1.0
  description: |
    API мини-приложения «Попутчики». Общие соглашения (аутентификация, ошибки, время, статусы,
    deep link) — в docs/contracts/README.md.
    `x-stage: pre-mvp` — нужно на первом этапе, `x-stage: mvp` — появится позже.
servers:
  - url: /api/v1
security:
  - telegramInitData: []
...
paths:
...
  /events/{event_id}/trips:
    parameters:
      - $ref: '#/components/parameters/EventId'
    get:
      operationId: listEventTrips
      tags: [trips]
      summary: Рейсы группы — для выбора в заявке (CAP-05)
      x-stage: pre-mvp
      responses:
        '200':
          description: Отсортированы по scheduled_arrival
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Trip' }
        '404': { $ref: '#/components/responses/NotFound' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '422': { $ref: '#/components/responses/ValidationError' }
    post:
      operationId: createTrip
      tags: [trips]
      summary: Добавить рейс вручную (только админ)
      x-stage: pre-mvp
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TripInput' }
      responses:
        '201':
          description: Создано
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Trip' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /trips/{trip_id}:
    parameters:
      - $ref: '#/components/parameters/TripId'
    patch:
      operationId: updateTrip
      tags: [trips]
      summary: Исправить рейс (только админ группы)
      description: Поля как в TripInput, все необязательны. Запись становится source=manual.
      x-stage: pre-mvp
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TripUpdate' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Trip' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    delete:
      operationId: deleteTrip
      tags: [trips]
      summary: Удалить рейс (только админ; нельзя, если на него ссылаются заявки)
      x-stage: pre-mvp
      responses:
        '204': { description: Удалено }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }   # TRIP_IN_USE
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationError' }
...
components:
  securitySchemes:
    telegramInitData:
      type: apiKey
      in: header
      name: Authorization
      description: '`tma <Telegram.WebApp.initData>`'
  parameters:
    EventId:
      name: event_id
      in: path
      required: true
      schema: { type: string, format: uuid }
    TripId:
      name: trip_id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    Unauthorized:
      description: UNAUTHORIZED
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: FORBIDDEN | PHONE_REQUIRED
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: NOT_FOUND
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    Conflict:
      description: Бизнес-конфликт, см. error.code
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    ValidationError:
      description: VALIDATION_ERROR, details.fields — ошибки по полям
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, details]
          properties:
            code:
              type: string
              enum:
                - VALIDATION_ERROR
                - UNAUTHORIZED
                - FORBIDDEN
                - PHONE_REQUIRED
                - NOT_FOUND
                - INVITE_NOT_FOUND
                - JOIN_CLOSED
                - APPLICATION_LOCKED
                - ALREADY_DEPARTED
                - CAPACITY_EXCEEDED
                - TRIP_IN_USE
                - INCOMPATIBLE_TRANSFER
                - FLIGHT_PROVIDER_UNAVAILABLE
                - INTERNAL_ERROR
            message: { type: string }
            details:
              type: [object, 'null']
              properties:
                fields:
                  type: object
                  additionalProperties: { type: string }
...
    LocationKind:
      type: string
      enum: [airport, station]
      description: station — задел под поезда, на текущих этапах не используется
    Location:
      type: object
      required: [id, kind, code, name, city, country, timezone]
      properties:
        id: { type: string, format: uuid }
        kind: { $ref: '#/components/schemas/LocationKind' }
        code: { type: string, description: IATA для аэропорта, example: SVO }
        name: { type: string, example: Шереметьево }
        city: { type: string, example: Москва }
        country: { type: string, description: ISO 3166-1 alpha-2, example: RU }
        timezone: { type: string, description: IANA, example: Europe/Moscow }
    TransportMode:
      type: string
      enum: [flight]
      description: Задел под train; сейчас только flight
    TripStatus:
      type: string
      enum: [scheduled, delayed, departed, landed, cancelled, unknown]
    TripInput:
      type: object
      required: [mode, number, departure_location_id, arrival_location_id,
                 scheduled_departure, scheduled_arrival]
      properties:
        mode: { $ref: '#/components/schemas/TransportMode' }
        number: { type: string, pattern: '^[A-Za-z0-9]{2,3}\s?\d{1,5}$', example: SU 1234 }
        departure_location_id: { type: string, format: uuid }
        arrival_location_id: { type: string, format: uuid }
        scheduled_departure:
          type: string
          format: date-time
          description: Без смещения — трактуется в таймзоне аэропорта вылета
        scheduled_arrival:
          type: string
          format: date-time
          description: Без смещения — трактуется в таймзоне аэропорта прилёта; > scheduled_departure
    TripUpdate:
      type: object
      properties:
        number: { type: string }
        departure_location_id: { type: string, format: uuid }
        arrival_location_id: { type: string, format: uuid }
        scheduled_departure: { type: string, format: date-time }
        scheduled_arrival: { type: string, format: date-time }
        estimated_arrival:
          type: [string, 'null']
          format: date-time
          description: Ручное исправление расчётного времени (в MVP может быть перезаписано поставщиком)
    Trip:
      type: object
      description: Транспортный рейс. Время отдаётся со смещением своего аэропорта
      required: [id, event_id, mode, number, departure_location, arrival_location,
                 scheduled_departure, scheduled_arrival, estimated_arrival, effective_arrival,
                 status, source, updated_at]
      properties:
        id: { type: string, format: uuid }
        event_id: { type: string, format: uuid }
        mode: { $ref: '#/components/schemas/TransportMode' }
        number: { type: string, example: SU1234, description: 'Нормализованный: без пробелов, верхний регистр' }
        departure_location: { $ref: '#/components/schemas/Location' }
        arrival_location: { $ref: '#/components/schemas/Location' }
        scheduled_departure: { type: string, format: date-time, example: '2026-11-12T07:30:00+05:00' }
        scheduled_arrival: { type: string, format: date-time, example: '2026-11-12T10:05:00+03:00' }
        estimated_arrival: { type: [string, 'null'], format: date-time }
        effective_arrival:
          type: string
          format: date-time
          description: estimated_arrival ?? scheduled_arrival — по нему идёт подбор
        status: { $ref: '#/components/schemas/TripStatus' }
        source: { type: string, enum: [manual, provider] }
        updated_at: { type: string, format: date-time }

    # ------------------------------------------------------------ заявка
...